Update to latest jupyter-widgets cookiecutter
Should now support jupyter lab
