+++ DEVELOPING.md
@@ -1,22 +1,25 @@
-# Development installation
+# Developing
 
-For a development installation (requires npm),
+For a development installation (requires [Node.js](https://nodejs.org) and [Yarn version 1](https://classic.yarnpkg.com/)),
 
     $ git clone https://github.com/ricklupton/ipysankeywidget.git
     $ cd ipysankeywidget
     $ pip install -e .
-    $ jupyter nbextension install --py --symlink --sys-prefix ipysankeywidget
-    $ jupyter nbextension enable --py --sys-prefix widgetsnbextension   # <-- you might have already done this,
+    $ jupyter nbextension install --py --symlink --overwrite --sys-prefix ipysankeywidget
     $ jupyter nbextension enable --py --sys-prefix ipysankeywidget
-    $ jupyter labextension install js @jupyter-widgets/jupyterlab-manager
-    
-When actively developing your extension, build Jupyter Lab with the command:
 
-    $ jupyter lab --watch
+Set up JS packages:
 
-This take a minute or so to get started, but then allows you to hot-reload your
-javascript extension. To see a change, save your javascript, watch the terminal
-for an update.
+    $ cd js
+    $ yarn install
 
-Note on first `jupyter lab --watch`, you may need to touch a file to get Jupyter
-Lab to open.
+When actively developing your extension for JupyterLab, run the command:
+
+    $ jupyter labextension develop --overwrite ipysankeywidget
+
+Then you need to rebuild the JS when you make a code change:
+
+    $ cd js
+    $ yarn run build
+
+You then need to refresh the JupyterLab page when your javascript changes.
