+++ install.json
@@ -1,0 +1,5 @@
+{
+  "packageManager": "python",
+  "packageName": "ipysankeywidget",
+  "uninstallInstructions": "Use your Python package manager (pip, conda, etc.) to uninstall the package ipysankeywidget"
+}
