+++ ipysankeywidget/_version.py
@@ -1,6 +1,4 @@
-version_info = (0, 4, 3, 'alpha', 0)
+# Module version
+__version__ = '0.5.0a0'
 
-_specifier_ = {'alpha': 'a', 'beta': 'b', 'candidate': 'rc', 'final': ''}
-
-__version__ = '%s.%s.%s%s'%(version_info[0], version_info[1], version_info[2],
-  '' if version_info[3]=='final' else _specifier_[version_info[3]]+str(version_info[4]))
+NPM_PACKAGE_RANGE='^0.5.0'
