+++ js/lib/index.js
@@ -1,3 +1,4 @@
 // Export widget models and views, and the npm package version number.
-module.exports = require('./sankey-widget.js');
-module.exports['version'] = require('../package.json').version;
+
+export {SankeyModel, SankeyView} from './sankey_widget';
+export {version} from '../package.json';
