--- postBuild
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-# This file is used by e.g. Binder to setup the environment
-
-# Enable extension for classic notebook
-jupyter nbextension enable --py --sys-prefix ipysankeywidget widgetsnbextension
-
-# Enable extension for Jupyterlab
-jupyter labextension install jupyter-sankey-widget @jupyter-widgets/jupyterlab-manager
