+++ setup.cfg
@@ -1,2 +1,35 @@
-[bdist_wheel]
-universal=1
+[metadata]
+name = ipysankeywidget
+author = Rick Lupton
+author_email=
+version = attr: ipysankeywidget._version.__version__
+description = Display Sankey diagrams in Jupyter
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/ricklupton/ipysankeywidget
+keywords = ipython, widgets, Jupyter, JupyterLab, JupyterLab3
+license = BSD 3-Clause License
+classifiers =
+    License :: OSI Approved :: BSD License
+    Development Status :: 4 - Beta
+    Framework :: IPython
+    Intended Audience :: Developers
+    Intended Audience :: Science/Research
+    Programming Language :: Python
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Framework :: Jupyter
+    Framework :: Jupyter :: JupyterLab
+    Framework :: Jupyter :: JupyterLab :: 3
+    Framework :: Jupyter :: JupyterLab :: Extensions
+    Framework :: Jupyter :: JupyterLab :: Extensions :: Prebuilt
+
+[options]
+zip_safe = False
+include_package_data = True
+packages = find:
+install_requires =
+    ipywidgets>=7.6.0,<9
+python_requires = >=3.7
